src/__tests__/index.test.ts
import { wrapRequest } from '../';

test('it should set loading state', () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(resolve, 0))
    );

    wrap.request();

    expect(wrap.fetched).toBeFalsy();
    expect(wrap.loading).toBeTruthy();
});

test('it should set fetched state', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(resolve, 0))
    );

    await wrap.request();

    expect(wrap.fetched).toBeTruthy();
    expect(wrap.loading).toBeFalsy();
});

test('it should set data', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(() => resolve(1337), 0))
    );

    const data = await wrap.request();

    expect(wrap.$).toEqual(1337);
    expect(data).toEqual(1337);
});

test('it should set error', async () => {
    const error = new Error('Error');
    const wrap = wrapRequest(
        () => new Promise((_, reject) => setTimeout(() => reject(error), 0))
    );

    await wrap.request();

    expect(wrap.error).toEqual(error);
});

test('it should match loading', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(() => resolve(1337), 0))
    );

    wrap.request();

    expect(
        wrap.match({
            loading: () => 'Loading',
            fetched: () => 'Fetched',
            empty: () => 'Empty',
            error: e => e
        })
    ).toEqual('Loading');
});

test('it should match fetched', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(() => resolve(1337), 0))
    );

    await wrap.request();

    expect(
        wrap.match({
            loading: () => 'Loading',
            fetched: () => 'Fetched',
            empty: () => 'Empty',
            error: e => e
        })
    ).toEqual('Fetched');
});

test('it should match empty', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(() => resolve([]), 0))
    );

    await wrap.request();

    expect(
        wrap.match({
            loading: () => 'Loading',
            fetched: () => 'Fetched',
            empty: () => 'Empty',
            error: e => e
        })
    ).toEqual('Empty');
});

test('it should match error', async () => {
    const wrap = wrapRequest(
        () => new Promise((_, reject) => setTimeout(() => reject('Error'), 0))
    );

    await wrap.request();

    expect(
        wrap.match({
            loading: () => 'Loading',
            fetched: () => 'Fetched',
            empty: () => 'Empty',
            error: e => e
        })
    ).toEqual('Error');
});

test('it should invoke ifFetched', async () => {
    const wrap = wrapRequest(
        () => new Promise(resolve => setTimeout(() => resolve(1337), 0))
    );

    await wrap.request();

    const fetchedValue = wrap.didFetch(val => `My val is ${val}`);

    expect(fetchedValue).toEqual('My val is 1337');
});

test('it should transform data', async () => {
    const wrap = wrapRequest(
        () =>
            new Promise<{ id: 1; name: 'Foo' }[]>(resolve =>
                setTimeout(() => resolve([{ id: 1, name: 'Foo' }]), 0)
            ),
        [],
        res => res[0]
    );

    const data = await wrap.request();

    expect(wrap.$).toEqual({ id: 1, name: 'Foo' });
    expect(data).toEqual({ id: 1, name: 'Foo' });
});
